WEB/requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.0
fsspec==2025.3.2
gradio==3.50.2
gradio_client==0.6.1
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.40.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
openai==1.79.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
ruff==0.11.10
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==11.0.3
Werkzeug==3.1.3